Control plane 2

+++ libs/idp-server-use-cases/src/main/java/org/idp/server/usecases/control_plane/ClientManagementEntryService.java
@@ -5,8 +5,9 @@ import java.util.List;
 import java.util.Map;
 import org.idp.server.basic.datasource.Transaction;
 import org.idp.server.basic.type.security.RequestAttributes;
+import org.idp.server.control_plane.base.definition.AdminPermissions;
 import org.idp.server.control_plane.management.oidc.client.*;
-import org.idp.server.control_plane.management.oidc.client.io.ClientConfigurationManagementResponse;
+import org.idp.server.control_plane.management.oidc.client.io.ClientManagementResponse;
 import org.idp.server.control_plane.management.oidc.client.io.ClientManagementStatus;
 import org.idp.server.control_plane.management.oidc.client.io.ClientRegistrationRequest;
 import org.idp.server.control_plane.management.oidc.client.validator.ClientRegistrationRequestValidationResult;
@@ -37,15 +38,29 @@ public class ClientManagementEntryService implements ClientManagementApi {
     this.clientConfigurationQueryRepository = clientConfigurationQueryRepository;
   }
 
-  public ClientConfigurationManagementResponse register(
+  public ClientManagementResponse create(
       TenantIdentifier tenantIdentifier,
       User operator,
       OAuthToken oAuthToken,
       ClientRegistrationRequest request,
-      RequestAttributes requestAttributes) {
+      RequestAttributes requestAttributes,
+      boolean dryRun) {
+
+    AdminPermissions permissions = getRequiredPermissions("create");
+    if (!permissions.includesAll(operator.permissionsAsSet())) {
+      Map<String, Object> response = new HashMap<>();
+      response.put("error", "access_denied");
+      response.put(
+          "error_description",
+          String.format(
+              "permission denied required permission %s, but %s",
+              permissions.valuesAsString(), operator.permissionsAsString()));
+      return new ClientManagementResponse(ClientManagementStatus.FORBIDDEN, response);
+    }
 
     Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
-    ClientRegistrationRequestValidator validator = new ClientRegistrationRequestValidator(request);
+    ClientRegistrationRequestValidator validator =
+        new ClientRegistrationRequestValidator(request, dryRun);
     ClientRegistrationRequestValidationResult validate = validator.validate();
 
     if (!validate.isValid()) {
@@ -53,30 +68,113 @@ public class ClientManagementEntryService implements ClientManagementApi {
     }
 
     ClientRegistrationContextCreator contextCreator =
-        new ClientRegistrationContextCreator(tenant, request);
+        new ClientRegistrationContextCreator(tenant, request, dryRun);
     ClientRegistrationContext context = contextCreator.create();
-    if (context.isDryRun()) {
+    if (dryRun) {
       return context.toResponse();
     }
 
-    clientConfigurationCommandRepository.register(tenant, context.clientConfiguration());
+    clientConfigurationCommandRepository.register(tenant, context.configuration());
 
     return context.toResponse();
   }
 
+  @Transaction(readOnly = true)
   @Override
-  public ClientConfigurationManagementResponse update(
+  public ClientManagementResponse findList(
+      TenantIdentifier tenantIdentifier,
+      User operator,
+      OAuthToken oAuthToken,
+      int limit,
+      int offset,
+      RequestAttributes requestAttributes) {
+
+    AdminPermissions permissions = getRequiredPermissions("findList");
+    if (!permissions.includesAll(operator.permissionsAsSet())) {
+      Map<String, Object> response = new HashMap<>();
+      response.put("error", "access_denied");
+      response.put(
+          "error_description",
+          String.format(
+              "permission denied required permission %s, but %s",
+              permissions.valuesAsString(), operator.permissionsAsString()));
+      return new ClientManagementResponse(ClientManagementStatus.FORBIDDEN, response);
+    }
+
+    Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
+
+    List<ClientConfiguration> clientConfigurations =
+        clientConfigurationQueryRepository.findList(tenant, limit, offset);
+    Map<String, Object> response = new HashMap<>();
+    response.put("list", clientConfigurations.stream().map(ClientConfiguration::toMap).toList());
+
+    return new ClientManagementResponse(ClientManagementStatus.OK, response);
+  }
+
+  @Transaction(readOnly = true)
+  @Override
+  public ClientManagementResponse get(
+      TenantIdentifier tenantIdentifier,
+      User operator,
+      OAuthToken oAuthToken,
+      ClientIdentifier clientIdentifier,
+      RequestAttributes requestAttributes) {
+
+    AdminPermissions permissions = getRequiredPermissions("get");
+    if (!permissions.includesAll(operator.permissionsAsSet())) {
+      Map<String, Object> response = new HashMap<>();
+      response.put("error", "access_denied");
+      response.put(
+          "error_description",
+          String.format(
+              "permission denied required permission %s, but %s",
+              permissions.valuesAsString(), operator.permissionsAsString()));
+      return new ClientManagementResponse(ClientManagementStatus.FORBIDDEN, response);
+    }
+
+    Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
+
+    ClientConfiguration clientConfiguration =
+        clientConfigurationQueryRepository.find(tenant, clientIdentifier);
+
+    if (!clientConfiguration.exists()) {
+      return new ClientManagementResponse(ClientManagementStatus.NOT_FOUND, Map.of());
+    }
+
+    return new ClientManagementResponse(ClientManagementStatus.OK, clientConfiguration.toMap());
+  }
+
+  @Override
+  public ClientManagementResponse update(
       TenantIdentifier tenantIdentifier,
       User operator,
       OAuthToken oAuthToken,
       ClientIdentifier clientIdentifier,
       ClientRegistrationRequest request,
-      RequestAttributes requestAttributes) {
+      RequestAttributes requestAttributes,
+      boolean dryRun) {
+
+    AdminPermissions permissions = getRequiredPermissions("update");
+    if (!permissions.includesAll(operator.permissionsAsSet())) {
+      Map<String, Object> response = new HashMap<>();
+      response.put("error", "access_denied");
+      response.put(
+          "error_description",
+          String.format(
+              "permission denied required permission %s, but %s",
+              permissions.valuesAsString(), operator.permissionsAsString()));
+      return new ClientManagementResponse(ClientManagementStatus.FORBIDDEN, response);
+    }
 
     Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
-    ClientConfiguration before = clientConfigurationQueryRepository.get(tenant, clientIdentifier);
+    ClientConfiguration before = clientConfigurationQueryRepository.find(tenant, clientIdentifier);
 
-    ClientRegistrationRequestValidator validator = new ClientRegistrationRequestValidator(request);
+    if (!before.exists()) {
+      return new ClientManagementResponse(ClientManagementStatus.NOT_FOUND, Map.of());
+    }
+
+    ClientRegistrationRequestValidator validator =
+        new ClientRegistrationRequestValidator(request, dryRun);
     ClientRegistrationRequestValidationResult validate = validator.validate();
 
     if (!validate.isValid()) {
@@ -84,9 +182,9 @@ public class ClientManagementEntryService implements ClientManagementApi {
     }
 
     ClientUpdateContextCreator contextCreator =
-        new ClientUpdateContextCreator(tenant, before, request);
+        new ClientUpdateContextCreator(tenant, before, request, dryRun);
     ClientUpdateContext context = contextCreator.create();
-    if (context.isDryRun()) {
+    if (dryRun) {
       return context.toResponse();
     }
 
@@ -96,56 +194,36 @@ public class ClientManagementEntryService implements ClientManagementApi {
   }
 
   @Override
-  public ClientConfigurationManagementResponse findList(
-      TenantIdentifier tenantIdentifier,
-      User operator,
-      OAuthToken oAuthToken,
-      int limit,
-      int offset,
-      RequestAttributes requestAttributes) {
-
-    Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
-
-    List<ClientConfiguration> clientConfigurations =
-        clientConfigurationQueryRepository.find(tenant, limit, offset);
-    Map<String, Object> response = new HashMap<>();
-    response.put("list", clientConfigurations.stream().map(ClientConfiguration::toMap).toList());
-
-    return new ClientConfigurationManagementResponse(ClientManagementStatus.OK, response);
-  }
-
-  @Override
-  public ClientConfigurationManagementResponse get(
+  public ClientManagementResponse delete(
       TenantIdentifier tenantIdentifier,
       User operator,
       OAuthToken oAuthToken,
       ClientIdentifier clientIdentifier,
-      RequestAttributes requestAttributes) {
+      RequestAttributes requestAttributes,
+      boolean dryRun) {
+
+    AdminPermissions permissions = getRequiredPermissions("delete");
+    if (!permissions.includesAll(operator.permissionsAsSet())) {
+      Map<String, Object> response = new HashMap<>();
+      response.put("error", "access_denied");
+      response.put(
+          "error_description",
+          String.format(
+              "permission denied required permission %s, but %s",
+              permissions.valuesAsString(), operator.permissionsAsString()));
+      return new ClientManagementResponse(ClientManagementStatus.FORBIDDEN, response);
+    }
 
     Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
-
     ClientConfiguration clientConfiguration =
-        clientConfigurationQueryRepository.get(tenant, clientIdentifier);
-    Map<String, Object> response = new HashMap<>();
-    response.put("client", clientConfiguration.toMap());
+        clientConfigurationQueryRepository.find(tenant, clientIdentifier);
 
-    return new ClientConfigurationManagementResponse(ClientManagementStatus.OK, response);
-  }
+    if (!clientConfiguration.exists()) {
+      return new ClientManagementResponse(ClientManagementStatus.NOT_FOUND, Map.of());
+    }
 
-  @Override
-  public ClientConfigurationManagementResponse delete(
-      TenantIdentifier tenantIdentifier,
-      User operator,
-      OAuthToken oAuthToken,
-      ClientIdentifier clientIdentifier,
-      RequestAttributes requestAttributes) {
-
-    Tenant tenant = tenantQueryRepository.get(tenantIdentifier);
-
-    ClientConfiguration clientConfiguration =
-        clientConfigurationQueryRepository.get(tenant, clientIdentifier);
     clientConfigurationCommandRepository.delete(tenant, clientConfiguration);
 
-    return new ClientConfigurationManagementResponse(ClientManagementStatus.NO_CONTENT, Map.of());
+    return new ClientManagementResponse(ClientManagementStatus.NO_CONTENT, Map.of());
   }
 }
